ml/requirements.txt
pandas
numpy
statsmodels
google-cloud-bigquery
python-dotenv
db-dtypes
prophet
yfinance
